requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Flask==2.3.2
Flask-Session==0.8.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
python-dotenv==1.0.1
Werkzeug==2.3.6
